% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.football.R
\name{draw.football}
\alias{draw.football}
\title{Draw an American football field}
\usage{
draw.football(length = 120, width = 53 + (1/3), main = "Football",
  col1 = "palegreen3", col2 = "white", col3 = "midnightblue")
}
\arguments{
\item{length}{the length (x-axis) of the field}

\item{width}{the width (y-axis) of the field}

\item{main}{the title that appears at the top of the plot}

\item{col1}{the color of the field itself}

\item{col2}{the color of the lines on the field}

\item{col3}{the color of the field in the endzones}
}
\value{
Plots a football field
}
\description{
Takes width, length, and color arguments to create a football field
}
\details{
This function creates a plot that acts as a base layer on which various
points and lines can be plotted.
}
\examples{
draw.football()
}
\author{
Braden Sharp
}
