% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.soccer.R
\name{draw.soccer}
\alias{draw.soccer}
\title{Draw a soccer field}
\usage{
draw.soccer(length = 120, width = 70, main = "Soccer",
  col1 = "palegreen3", col2 = "white")
}
\arguments{
\item{length}{the length (x-axis) of the field}

\item{width}{the width (y-axis) of the field}

\item{main}{the title that appears at the top of the plot}

\item{col1}{the color of the field itself}

\item{col2}{the color of the boundary lines on the field}
}
\value{
Plots a soccer field
}
\description{
Takes width, length, and color arguments to create a soccer field
}
\details{
This function creates a plot that acts as a base layer on which various
points and lines can be plotted.
}
\examples{
draw.soccer()
}
\author{
Braden Sharp
}
