% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.tennis.R
\name{draw.tennis}
\alias{draw.tennis}
\title{Draw a tennis court}
\usage{
draw.tennis(length = 78, width = 36, main = "Tennis",
  col1 = "aquamarine3", col2 = "white")
}
\arguments{
\item{length}{the length (x-axis) of the court}

\item{width}{the width (y-axis) of the court}

\item{main}{the title that appears at the top of the plot}

\item{col1}{the color of the court itself}

\item{col2}{the color of the lines on the court}
}
\value{
Plots a tennis court
}
\description{
Takes width, length, and color arguments to create a tennis court
}
\details{
This function creates a plot that acts as a base layer on which various
points and lines can be plotted.
}
\examples{
draw.tennis()
}
\author{
Braden Sharp
}
