% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.volleyball.R
\name{draw.volleyball}
\alias{draw.volleyball}
\title{Draw a volleyball court}
\usage{
draw.volleyball(length = 18, width = 9, main = "Volleyball",
  col1 = "burlywood2", col2 = "darkblue")
}
\arguments{
\item{length}{the length (x-axis) of the court}

\item{width}{the width (y-axis) of the court}

\item{main}{the title that appears at the top of the plot}

\item{col1}{the color of the court itself}

\item{col2}{the color of the lines on the court}
}
\value{
Plots a volleyball court
}
\description{
Takes width, length, and color arguments to create a volleyball court
}
\details{
This function creates a plot that acts as a base layer on which various
points and lines can be plotted.
}
\examples{
draw.volleyball()
}
\author{
Braden Sharp
}
